# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Epup Index Creator
    version: 0.0.1
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /books:
        get:
            tags:
                - epub_index_creator
            summary: ListBooks epub_index_creator
            operationId: epub_index_creator#ListBooks
            parameters:
                - name: ListBooksRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/EpubIndexCreatorListBooksRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/BookResponse'
            schemes:
                - http
    /books/{isbn}:
        get:
            tags:
                - epub_index_creator
            summary: FindBook epub_index_creator
            operationId: epub_index_creator#FindBook
            parameters:
                - name: isbn
                  in: path
                  description: ISBN of the book
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/EpubIndexCreatorFindBookResponseBody'
                        required:
                            - isbn
                            - title
                            - author
                            - pages
            schemes:
                - http
    /openapi.json:
        get:
            tags:
                - epub_index_creator
            summary: Download ./gen/http/openapi.json
            operationId: epub_index_creator#/openapi.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
definitions:
    BookResponse:
        title: BookResponse
        type: object
        properties:
            author:
                type: string
                description: Author of the book
                example: Dolor provident nulla corporis aspernatur.
            isbn:
                type: string
                description: ISBN of the book
                example: Atque labore ad.
            pages:
                type: array
                items:
                    $ref: '#/definitions/PageResponse'
                description: Pages of the book
                example:
                    - keywords:
                        - keyword: Non consequatur dolore ad tempore aut.
                        - keyword: Non consequatur dolore ad tempore aut.
                        - keyword: Non consequatur dolore ad tempore aut.
                        - keyword: Non consequatur dolore ad tempore aut.
                      title: In omnis quaerat odit numquam ut repudiandae.
                    - keywords:
                        - keyword: Non consequatur dolore ad tempore aut.
                        - keyword: Non consequatur dolore ad tempore aut.
                        - keyword: Non consequatur dolore ad tempore aut.
                        - keyword: Non consequatur dolore ad tempore aut.
                      title: In omnis quaerat odit numquam ut repudiandae.
            title:
                type: string
                description: Title of the book
                example: Laboriosam quae consequatur autem autem facilis aliquam.
        example:
            author: Harum quia.
            isbn: Quasi molestiae veniam.
            pages:
                - keywords:
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                  title: In omnis quaerat odit numquam ut repudiandae.
                - keywords:
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                  title: In omnis quaerat odit numquam ut repudiandae.
                - keywords:
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                  title: In omnis quaerat odit numquam ut repudiandae.
                - keywords:
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                  title: In omnis quaerat odit numquam ut repudiandae.
            title: Est sit error voluptatum asperiores.
        required:
            - isbn
            - title
            - author
            - pages
    EpubIndexCreatorFindBookResponseBody:
        title: EpubIndexCreatorFindBookResponseBody
        type: object
        properties:
            author:
                type: string
                description: Author of the book
                example: Sit quia neque ratione quidem.
            isbn:
                type: string
                description: ISBN of the book
                example: Non consequatur.
            pages:
                type: array
                items:
                    $ref: '#/definitions/PageResponseBody'
                description: Pages of the book
                example:
                    - keywords:
                        - keyword: Veniam est aut occaecati odio corrupti quis.
                        - keyword: Veniam est aut occaecati odio corrupti quis.
                        - keyword: Veniam est aut occaecati odio corrupti quis.
                      title: Velit dolores aperiam adipisci qui.
                    - keywords:
                        - keyword: Veniam est aut occaecati odio corrupti quis.
                        - keyword: Veniam est aut occaecati odio corrupti quis.
                        - keyword: Veniam est aut occaecati odio corrupti quis.
                      title: Velit dolores aperiam adipisci qui.
            title:
                type: string
                description: Title of the book
                example: Voluptas est dolorem et.
        example:
            author: Aut et exercitationem placeat ad.
            isbn: Voluptatem corporis ex dignissimos sed qui.
            pages:
                - keywords:
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                  title: Velit dolores aperiam adipisci qui.
                - keywords:
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                  title: Velit dolores aperiam adipisci qui.
                - keywords:
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                  title: Velit dolores aperiam adipisci qui.
            title: Consequatur aspernatur sint quam nemo.
        required:
            - isbn
            - title
            - author
            - pages
    EpubIndexCreatorListBooksRequestBody:
        title: EpubIndexCreatorListBooksRequestBody
        type: object
        properties:
            limit:
                type: integer
                description: Maximum number of books to return
                default: 100
                example: 1
                format: int64
                minimum: 1
                maximum: 100
            offset:
                type: integer
                description: Field to paginate books
                default: 0
                example: 7026938516198184411
                format: int64
                minimum: 0
        example:
            limit: 11
            offset: 2784089440402910752
    KeywordResponse:
        title: KeywordResponse
        type: object
        properties:
            keyword:
                type: string
                description: Keyword of the page
                example: Perferendis voluptatibus aliquam.
        example:
            keyword: Molestias aliquam sunt et numquam esse in.
        required:
            - keyword
    KeywordResponseBody:
        title: KeywordResponseBody
        type: object
        properties:
            keyword:
                type: string
                description: Keyword of the page
                example: Est ex asperiores aspernatur enim mollitia laboriosam.
        example:
            keyword: Tenetur iusto animi mollitia eius magni.
        required:
            - keyword
    PageResponse:
        title: PageResponse
        type: object
        properties:
            keywords:
                type: array
                items:
                    $ref: '#/definitions/KeywordResponse'
                description: Keywords of the page
                example:
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
                    - keyword: Non consequatur dolore ad tempore aut.
            title:
                type: string
                description: Title of the page
                example: Dicta accusantium officiis totam omnis qui.
        example:
            keywords:
                - keyword: Non consequatur dolore ad tempore aut.
                - keyword: Non consequatur dolore ad tempore aut.
            title: Aperiam in reiciendis quaerat temporibus omnis.
        required:
            - title
            - keywords
    PageResponseBody:
        title: PageResponseBody
        type: object
        properties:
            keywords:
                type: array
                items:
                    $ref: '#/definitions/KeywordResponseBody'
                description: Keywords of the page
                example:
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
                    - keyword: Veniam est aut occaecati odio corrupti quis.
            title:
                type: string
                description: Title of the page
                example: Autem animi.
        example:
            keywords:
                - keyword: Veniam est aut occaecati odio corrupti quis.
                - keyword: Veniam est aut occaecati odio corrupti quis.
                - keyword: Veniam est aut occaecati odio corrupti quis.
            title: Omnis qui.
        required:
            - title
            - keywords